feat(graphic): add OH_Drawing_GetFontPathsByType API
Introduce a new API to obtain an array of font paths from the system based on
the specified font type. This function returns a pointer to an array of
OH_Drawing_String representing the font paths, and updates the path count via
output parameter.

The new API is also registered in the NDK JSON manifest for version 23.

Signed-off-by: liufei <liufei225@h-partners.com>